refactor(app): move app.py -> app/app.py and demo-app.py -> app.py
